requirements.txt
flake8==5.0.4
flake8-annotations==2.9.1
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
pep8-naming==0.13.2
pytest==7.1.3
pandas==2.0.3
mysql-connector-python==9.0.0
openpyxl==3.1.5
